% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainDataOutput.R
\name{mod_mainData}
\alias{mod_mainData}
\title{mod_mainData server function}
\usage{
mod_mainData(
  input,
  output,
  session,
  data_reactives,
  map_reactives,
  modosindb,
  lang
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives, map_reactives}{reactives from modules}

\item{lang}{lang selected}

\item{catdroughtdb}{object to access the meteoland db}
}
\description{
mod_mainData server function
}
\details{
mod_mainData always return the data in the 3043 projection
}
