% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{modosin_data}
\alias{modosin_data}
\title{modosin_data server function}
\usage{
modosin_data(input, output, session, modosindb, lang)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}
}
\description{
modosin_data server function
}
